requirements.txt
streamlit
pandas
yfinance
matplotlib
numpy
hurst
Pillow
